Fix crash when using compare bills feature
The compare bills feature was crashing, this was due to changes in
introduced in the compare bill numbering code when Pandas was
removed. Before the refactor to remove Pandas there was an to_html
method which was inadvertently removed. A new to_html_tables has
been added and this required minor changes to the compare bills