Further change to loading main UI html and css etc
Further to fd9261a5, in an attempt to use pywebview witout an HTTP server (bottle is the default). I have changed the way the html, css and js for the main UI is bundeled. Now we are creating a single .html file for the ui (that contains all the CSS and JS / React). This was because on windows I was getting an CORS related issue when trying to load the css and js. This was not a problem on Mac OS so long as the CSS and JS files were in the same folder as the HTML or a subfolder...

We now need the vite plugin, 'vite-plugin-singlefile' to bundle the html, css and js into a single file. Changes to the vite.config.ts file are required for this. Bootstrap is now installed via NPM.